+++ jovo-platforms/jovo-platform-sapcai/package.json
@@ -7,6 +7,7 @@
   "scripts": {
     "tslint": "node_modules/.bin/tslint -p tsconfig.json -c tslint.json",
     "tsc": "node_modules/.bin/tsc",
+    "watch": "tsc -w",
     "rimraf": "node_modules/.bin/rimraf ./dist",
     "typedoc": "node_modules/.bin/typedoc --out ./docs ./ --mode file",
     "test": "node_modules/.bin/jest"
