+++ jovo-platforms/jovo-platform-sapcai/sample-request-json/v1/IntentRequest.json
