+++ jovo-platforms/jovo-platform-sapcai/src/core/SapCaiResponse.ts
@@ -1,0 +1,352 @@
+import { JovoResponse, SessionConstants, SessionData } from 'jovo-core';
+import { CardContent } from '../response';
+import { Message } from './Interfaces';
+import _get = require('lodash.get');
+import _set = require('lodash.set');
+
+export interface SessionAttributes {
+  [key: string]: any; //tslint:disable-line
+}
+
+export interface SAPCAIResponseJSON {
+  replies?: any[];
+  conversation?: ConversationData;
+}
+
+export interface ConversationData {
+  memory?: SessionAttributes;
+}
+
+export class SapCaiResponse implements JovoResponse {
+  replies: Message[];
+  conversation?: ConversationData;
+
+  constructor() {
+    this.replies = [];
+    this.conversation = { memory: {} };
+  }
+
+  getFirstReply(type: string) {
+    const items: Message[] = _get(this, 'replies');
+
+    for (let i = 0; i < items.length; i++) {
+      if (items[i] && items[i].type === type) {
+        return items[i];
+      }
+    }
+  }
+
+  getBasicText() {
+    return this.getFirstReply('text');
+  }
+
+  getBasicCard() {
+    return this.getFirstReply('card');
+  }
+
+  getQuickReplyCard() {
+    return this.getFirstReply('quickReplies');
+  }
+
+  hasQuickReplyCard(title: string, ...chips: string[]) {
+    const quickReplies = this.getQuickReplyCard();
+
+    // TODO cleanup
+
+    if (
+      !quickReplies ||
+      !title ||
+      typeof quickReplies.content === 'string' ||
+      !quickReplies.content.title ||
+      quickReplies.content.title !== title ||
+      !quickReplies.content.buttons
+    ) {
+      return false;
+    }
+
+    for (let i = 0; i < chips.length; i++) {
+      if (!quickReplies.content.buttons[i] || chips[i] !== quickReplies.content.buttons[i].title) {
+        return false;
+      }
+    }
+
+    return true;
+  }
+
+  getButtonsCard() {
+    return this.getFirstReply('buttons');
+  }
+
+  hasButtonsCard(title: string, ...chips: string[]) {
+    const buttonList = this.getButtonsCard();
+
+    if (
+      !buttonList ||
+      !title ||
+      typeof buttonList.content === 'string' ||
+      !buttonList.content.title ||
+      buttonList.content.title !== title ||
+      !buttonList.content.buttons
+    ) {
+      return false;
+    }
+
+    for (let i = 0; i < chips.length; i++) {
+      if (!buttonList.content.buttons[i] || chips[i] !== buttonList.content.buttons[i].title) {
+        return false;
+      }
+    }
+
+    return true;
+  }
+
+  getCarouselCard() {
+    return this.getFirstReply('carousel');
+  }
+
+  hasCarouselCard(...items: CardContent[]) {
+    const carousel = this.getCarouselCard();
+
+    // TODO cleanup
+    if (!carousel || typeof carousel.content === 'string' || !Array.isArray(carousel.content)) {
+      return false;
+    }
+
+    for (let i = 0; i < items.length; i++) {
+      if (!carousel.content[i] || items[i].title !== carousel.content[i].title) {
+        return false;
+      }
+      if (!carousel.content[i] || items[i].subtitle !== carousel.content[i].subtitle) {
+        return false;
+      }
+      if (!carousel.content[i] || items[i].imageUrl !== carousel.content[i].imageUrl) {
+        return false;
+      }
+      if (items[i].buttons) {
+        for (let j = 0; j < items[i].buttons!.length; j++) {
+          if (
+            !carousel.content[i].buttons ||
+            items[i].buttons![j].title !== carousel.content[i].buttons[j].title
+          ) {
+            return false;
+          }
+        }
+      }
+    }
+
+    return true;
+  }
+
+  getListCard() {
+    return this.getFirstReply('list');
+  }
+
+  hasListCard(elements: CardContent[], buttonChips: string[]) {
+    const list = this.getListCard();
+
+    // TODO cleanup
+    if (!list) {
+      return false;
+    }
+
+    for (let i = 0; i < elements.length; i++) {
+      if (!list.content.elements[i] || elements[i].title !== list.content.elements[i].title) {
+        return false;
+      }
+      if (!list.content.elements[i] || elements[i].subtitle !== list.content.elements[i].subtitle) {
+        return false;
+      }
+      if (!list.content.elements[i] || elements[i].imageUrl !== list.content.elements[i].imageUrl) {
+        return false;
+      }
+      if (elements[i].buttons !== undefined) {
+        for (let j = 0; j < elements[i].buttons!.length; j++) {
+          if (
+            !list.content.elements[i].buttons ||
+            elements[i].buttons![j].title !== list.content.elements[i].buttons[j].title
+          ) {
+            return false;
+          }
+        }
+      }
+    }
+
+    if (buttonChips) {
+      for (let i = 0; i < buttonChips.length; i++) {
+        if (!list.content.buttons[i] || buttonChips[i] !== list.content.buttons[i].title) {
+          return false;
+        }
+      }
+    }
+
+    return true;
+  }
+
+  getImageCard() {
+    return this.getFirstReply('picture');
+  }
+
+  hasImageCard(pictureUrl: string) {
+    const card = this.getImageCard();
+
+    if (!card) {
+      return false;
+    }
+
+    if (!pictureUrl || !card.content || card.content !== pictureUrl) {
+      return false;
+    }
+
+    return true;
+  }
+
+  hasStandardCard(title?: string, subtitle?: string, imageUrl?: string): boolean {
+    const basicCardObject = this.getBasicCard();
+
+    if (!basicCardObject) {
+      return false;
+    }
+
+    if (title) {
+      if (title !== basicCardObject.content.title) {
+        return false;
+      }
+    }
+
+    if (subtitle) {
+      if (subtitle !== basicCardObject.content.subtitle) {
+        return false;
+      }
+    }
+
+    if (imageUrl) {
+      if (imageUrl !== basicCardObject.content.imageUrl) {
+        return false;
+      }
+    }
+
+    return true;
+  }
+
+  getSessionData(path?: string) {
+    if (path) {
+      return this.getSessionAttribute(path);
+    } else {
+      return this.getSessionAttributes();
+    }
+  }
+
+  hasSessionData(name: string, value?: any): boolean {
+    // tslint:disable-line
+    return this.hasSessionAttribute(name, value);
+  }
+
+  setSessionData(sessionData: SessionData) {
+    return this.setSessionAttributes(sessionData);
+  }
+
+  getSessionAttributes() {
+    return _get(this, 'conversation.memory');
+  }
+
+  setSessionAttributes(sessionData: SessionData) {
+    _set(this, 'conversation.memory', sessionData);
+    return this;
+  }
+
+  getSpeech() {
+    return undefined;
+  }
+
+  getReprompt() {
+    return undefined;
+  }
+
+  getSpeechPlain() {
+    return undefined;
+  }
+
+  getRepromptPlain() {
+    return undefined;
+  }
+
+  getSessionAttribute(name: string) {
+    return _get(this, `conversation.memory.${name}`);
+  }
+
+  hasSessionAttribute(name: string, value?: any): boolean {
+    // tslint:disable-line
+    if (!this.getSessionAttribute(name)) {
+      return false;
+    }
+
+    if (typeof value !== 'undefined') {
+      if (this.getSessionAttribute(name) !== value) {
+        return false;
+      }
+    }
+
+    return true;
+  }
+
+  hasState(state: string): boolean {
+    return this.hasSessionAttribute(SessionConstants.STATE, state);
+  }
+
+  hasSessionEnded() {
+    return true;
+  }
+
+  /**
+   * Checks if response is a tell request
+   * @param {string| string[]} speechText
+   * @return {boolean}
+   */
+  isTell(speechText?: string | string[]): boolean {
+    if (speechText) {
+      const ssml: string = this.getBasicText().content;
+
+      if (Array.isArray(speechText)) {
+        for (const speechTextElement of speechText) {
+          if (speechTextElement === ssml) {
+            return true;
+          }
+        }
+        return false;
+      } else {
+        return ssml === speechText;
+      }
+    }
+    return true;
+  }
+
+  isAsk(speechText?: string | string[], repromptText?: string | string[]): boolean {
+    return false;
+  }
+
+  toJSON(): SAPCAIResponseJSON {
+    // copy all fields from `this` to an empty object and return in
+    return Object.assign({}, this);
+  }
+
+  // fromJSON is used to convert an serialized version
+  // of the User to an instance of the class
+  static fromJSON(json: SAPCAIResponseJSON | string) {
+    if (typeof json === 'string') {
+      // if it's a string, parse it first
+      return JSON.parse(json, SapCaiResponse.reviver);
+    } else {
+      // create an instance of the User class
+      const sapcaiResponse = Object.create(SapCaiResponse.prototype);
+      // copy all the fields from the json object
+      return Object.assign(sapcaiResponse, json);
+    }
+  }
+
+  // reviver can be passed as the second parameter to JSON.parse
+  // to automatically call User.fromJSON on the resulting value.
+  static reviver(key: string, value: any): any {
+    // tslint:disable-line
+    return key === '' ? SapCaiResponse.fromJSON(value) : value;
+  }
+}
