+++ jovo-platforms/jovo-platform-sapcai/src/core/SapCaiResponseBuilder.ts
@@ -1,0 +1,9 @@
+import { SapCaiResponse } from './SapCaiResponse';
+import { ResponseBuilder } from 'jovo-core';
+
+export class SapCaiResponseBuilder implements ResponseBuilder<SapCaiResponse> {
+  create(json: any): SapCaiResponse {
+    // tslint:disable-line
+    return SapCaiResponse.fromJSON(json);
+  }
+}
