+++ jovo-platforms/jovo-platform-sapcai/src/response/BasicCard.ts
@@ -1,17 +1,16 @@
-
 export interface BasicCard {
-    type: string;
+  type: string;
 }
 
 export abstract class BasicCard {
-    type: string;
+  type: string;
 
-    constructor(type: string) {
-        this.type = type;
-    }
+  constructor(type: string) {
+    this.type = type;
+  }
 
-    setType(type: string): this {
-        this.type = type;
-        return this;
-    }
+  setType(type: string): this {
+    this.type = type;
+    return this;
+  }
 }
