+++ jovo-platforms/jovo-platform-sapcai/src/response/Button.ts
@@ -1,14 +1,13 @@
-
 import _merge = require('lodash.merge');
 
 export class Button {
-    title?: string;
-    value?: string;
-    type?: string;
+  title?: string;
+  value?: string;
+  type?: string;
 
-    constructor(title: string, value: string, type?: string) {
-        this.title = title;
-        this.value = value;
-        this.type = type;
-    }
+  constructor(title: string, value: string, type?: string) {
+    this.title = title;
+    this.value = value;
+    this.type = type;
+  }
 }
