+++ jovo-platforms/jovo-platform-sapcai/src/response/ButtonList.ts
@@ -1,12 +1,11 @@
 import { BasicCard } from './BasicCard';
 import { CardContent } from './Card';
 
-
 export class ButtonList extends BasicCard {
-    content?: CardContent;
+  content?: CardContent;
 
-    constructor(content: CardContent) {
-        super('buttons');
-        this.content = content;
-    }
+  constructor(content: CardContent) {
+    super('buttons');
+    this.content = content;
+  }
 }
