+++ jovo-platforms/jovo-platform-sapcai/src/response/Card.ts
@@ -1,18 +1,18 @@
 import { BasicCard } from './BasicCard';
-import { Button } from "./Button";
+import { Button } from './Button';
 
 export class CardContent {
-    title?: string;
-    subtitle?: string;
-    imageUrl?: string;
-    buttons?: Button[];
+  title?: string;
+  subtitle?: string;
+  imageUrl?: string;
+  buttons?: Button[];
 }
 
 export class Card extends BasicCard {
-    content?: CardContent;
+  content?: CardContent;
 
-    constructor(content: CardContent) {
-        super('card');
-        this.content = content;
-    }
+  constructor(content: CardContent) {
+    super('card');
+    this.content = content;
+  }
 }
