+++ jovo-platforms/jovo-platform-sapcai/src/response/Carousel.ts
@@ -1,12 +1,11 @@
 import { BasicCard } from './BasicCard';
 import { CardContent } from './Card';
 
-
 export class Carousel extends BasicCard {
-    content?: CardContent[];
+  content?: CardContent[];
 
-    constructor(items: CardContent[]) {
-        super('carousel');
-        this.content = items;
-    }
+  constructor(items: CardContent[]) {
+    super('carousel');
+    this.content = items;
+  }
 }
