+++ jovo-platforms/jovo-platform-sapcai/src/response/List.ts
@@ -3,20 +3,20 @@ import { CardContent } from './Card';
 import { Button } from './Button';
 
 export class ListContent {
-    elements?: CardContent[];
-    buttons?: Button[];
+  elements?: CardContent[];
+  buttons?: Button[];
 
-    constructor(elements: CardContent[], buttons: Button[]) {
-        this.elements = elements;
-        this.buttons = buttons;
-    }
+  constructor(elements: CardContent[], buttons: Button[]) {
+    this.elements = elements;
+    this.buttons = buttons;
+  }
 }
 
 export class List extends BasicCard {
-    content?: ListContent;
+  content?: ListContent;
 
-    constructor(content: ListContent) {
-        super('list');
-        this.content = content;
-    }
+  constructor(content: ListContent) {
+    super('list');
+    this.content = content;
+  }
 }
